lanes: counts are u32, floats are bitcast — the 2^24 wall falls
The slot buffer held twelve lanes of two different kinds in one f32 array.
Five are integer COUNTS (S_ROW/S_COL/S_FLAGS/S_ORD, plus S_GLYPH_ID);
seven are genuine floats. f32 represents consecutive integers exactly only
to 2^24, so the counts capped the WHOLE arena at 16MB — and past it the
failure was silent: two glyphs collapse onto one ordinal while addressing
stays perfectly exact. Nothing throws; the layout is just wrong.

The buffer is u32 now. Counts store natively and are exact across the full
range — that is the entire point. Floats bitcast, which is the direction
that costs nothing: a reinterpretation preserves all 32 bits and round-
trips exactly, unlike the int-in-float it replaces, which destroyed
information. fbits()/fval() on the CPU stage through a Float32Array, so
every store rounds exactly where it rounded before; TSL bitcast() on the
GPU, a primitive these kernels already used for the atomic fold scalars.

S_GLYPH_ID is deliberately classified FLOAT: it is copied straight from
the trie's f32 blocks, so it stays a float carrier until the trie format
itself moves. Getting that wrong once — raw in the oracle, bitcast in the
kernels — made oracle and GPU disagree on every glyph id.

THE CEILING, now collected. KERNEL_MAX_BYTES becomes 2^32/SLOT_STRIDE
(341MB, 21x), bounded by the u32 slot INDEX rather than lane precision.
Three things had to be true first, and two were not:
  - glyphVertex converted a natively-unsigned instanceIndex DOWN to i32,
    wrapping at 2^31/STRIDE — half any u32 bound, silently, with every
    glyph past it reading a negative index. Now u32, matching the kernels.
  - Nothing validated the allocation against the device. assertSlotBufferFits
    refuses at the seam naming both numbers, instead of dying inside Dawn
    with no mention of glyphs or arenas. Advisory when limits are unreachable.
  - stage() called _alloc (which mutates) before growth, so a _realloc
    throw escaped WITHOUT returning the range — an unowned range poisons
    the watermark permanently and every later stage re-enters the throw.
    Now try/_insertFree/rethrow: refusals stay recoverable, per file.
The arena IMPORTS the kernels' constant rather than restating it. Two
copies of a limit is how this ceiling once got raised on one side only.

LANE_KIND is the root-cause fix. The counts-vs-floats classification lived
in prose and in ad-hoc sets copied into gen.mjs, verifyItem and four
comparators — and every copy drifted at least once: verifyItem carried
hardcoded indices stale by one since the codepoint lane fell (13 -> 12),
three tolerance comparators silently measured BIT PATTERNS (a 1e-4 gate on
bits is ~1.3% on values, ~100x looser than intended), and tools/lanemap
held a table classifying S_GLYPH_ID as a count, contradicting the others.
One exported table now; consumers derive; a totality guard fails if a lane
is ever added without classifying it.

Found by review, not by me: the storage NODE that binds this buffer in the
vertex and pick shaders was still declared 'float'. WGSL binds the same
memory as array<f32> with no validation error, so .toFloat() was a no-op
and vRowCol received bit patterns — row 5 as 0x5, a denormal, ~0 — which
would have shipped every glyph's far-LOD UV sampling texel (0,0). Near
text is unaffected, which is exactly why 47 test files, five kernel modes
and a byte-identical corpus all stayed green over it.

VERIFICATION. The engine fixture corpus regenerates BYTE-IDENTICAL
(29529386797a703c) from the migrated oracle: representation moved, no
value did. That check caught three bugs the test suite could not see.
tools/arena-capacity.test.mjs is new and closes the gap that let a
half-landed ceiling pass everything: every other test here runs a SMALL
arena and none exercise capacity at all. It is mutation-tested against
both real bugs — a divergent ceiling trips two assertions, a leaked range
shows live 4096 where 0 was expected.

48/48 JS files, four GPU gates clean, app build clean, smoke boots with
zero errors.

+++ packages/glyph3d-core/src/GlyphField.js
@@ -724,8 +724,15 @@ function _fieldUniform(initial, read) {
  * attaches. Read-only: vertex-stage storage is core WebGPU only as read-only.
  */
 function _fieldSlots() {
-    const placeholder = new StorageInstancedBufferAttribute(new Float32Array(4), 1);
-    return registerByteSlotsNode(storage(placeholder, 'float', 1).toReadOnly().onObjectUpdate(({ object }, self) => {
+    // 'uint', matching the slot buffer's own type (glyphPipelineKernels: the
+    // slots array is instancedArray(..., 'uint')). This declaration is what the
+    // shader reads the memory AS — WGSL will bind the same bytes as array<f32>
+    // with no validation error, so a stale 'float' here silently reinterprets
+    // every count lane as a denormal (row 5 -> 0x5 -> ~7e-45 -> 0) and makes
+    // glyphVertex's .toFloat() a no-op. Near text is unaffected; the far-LOD UV
+    // collapses to texel (0,0) for every glyph. Nothing errors.
+    const placeholder = new StorageInstancedBufferAttribute(new Uint32Array(4), 1);
+    return registerByteSlotsNode(storage(placeholder, 'uint', 1).toReadOnly().onObjectUpdate(({ object }, self) => {
         const f = object && object.userData && object.userData.glyphField;
         return (f && f._byteSlots) || self.value;
     }));
